ng2: remove ng1 components from account
